requirements.txt
# Weather data pipeline requirements
requests>=2.31.0
pandas>=2.0.0
PyYAML>=6.0
SQLAlchemy>=2.0.0

# Dashboard requirements
streamlit>=1.28.0
plotly>=5.15.0
folium>=0.15.0
streamlit-folium>=0.15.0

# ML requirements
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# SMS notification requirements
twilio 